# openapi.json
{
...
  "info": {
    "contact": {
      "email": "tarsom27v@gmail.com",
      "name": "Tarso Maia"
    },
    "description": "Um Encurtador de URL desenvolvido com Flask",
    "title": "Encurtador de URL",
    "version": "1"
  },
  "paths": {
...
    "/users": {
      "post": {
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "email": {
                  "example": "joao@email.com",
                  "type": "string"
                },
                "password": {
                  "example": "senha123",
                  "format": "password",
                  "type": "string"
                },
                "username": {
                  "example": "joao_silva",
                  "type": "string"
                }
              },
              "required": [
                "username",
                "password",
                "email"
              ],
              "type": "object"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Usu\u00e1rio criado com sucesso",
            "schema": {
              "properties": {
                "msg": {
                  "example": "Usu\u00e1rio criado",
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "400": {
            "description": "Dados inv\u00e1lidos ou ausentes",
            "schema": {
              "properties": {
                "msg": {
                  "enum": [
                    "JSON inv\u00e1lido",
                    "Dados faltando",
                    "Email inv\u00e1lido"
                  ],
                  "example": "Dados faltando",
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "409": {
            "description": "Usu\u00e1rio ou email j\u00e1 existente",
            "schema": {
              "properties": {
                "msg": {
                  "example": "Usu\u00e1rio ou email j\u00e1 existente",
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "500": {
            "description": "Erro interno",
            "schema": {
              "properties": {
                "msg": {
                  "enum": [
                    "Erro no banco de dados",
                    "Erro inesperado"
                  ],
                  "example": "Erro inesperado",
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        },
        "summary": "Registra um novo usu\u00e1rio",
        "tags": [
          "Usu\u00e1rios"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "deprecated": true,
        "parameters": [
          {
            "description": "ID do usu\u00e1rio",
            "example": 1,
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "201": {
            "description": "Usu\u00e1rio encontrado",
            "schema": {
              "properties": {
                "email": {
                  "example": "joao@email.com",
                  "type": "string"
                },
                "username": {
                  "example": "joao_silva",
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "404": {
            "description": "Usu\u00e1rio n\u00e3o encontrado"
          }
        },
        "summary": "Busca um usu\u00e1rio pelo ID",
        "tags": [
          "Usu\u00e1rios"
        ]
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "description": "Informe: Bearer <seu_token>",
      "in": "header",
      "name": "Authorization",
      "type": "apiKey"
    }
  },
  "swagger": "2.0"
}